Azure Chat Bot/requirements.txt
botbuilder-core==4.15.0
botbuilder-dialogs==4.15.0
requests
pytest
cryptography
